Move domain things to domain
