+++ libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/components/EventDetailComponent.kt
