+++ libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/components/ExpenseDetailComponent.kt
@@ -21,4 +21,3 @@ interface ExpenseDetailComponent {
         val venue: Venue? = null,
     )
 }
-
