+++ libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/components/InsertExpenseComponent.kt
@@ -41,4 +41,3 @@ interface InsertExpenseComponent {
         val isSubmitEnabled: Boolean = true,
     )
 }
-
