+++ libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/components/RootComponent.kt
@@ -3,6 +3,10 @@ package com.github.saikcaskey.pokertracker.domain.components
 import com.arkivanov.decompose.router.stack.ChildStack
 import com.arkivanov.decompose.value.Value
 
+/**
+ * The navigation "Root" for the app's decompose structure.
+ * All of these children can be pushed onto the Root navigator's Stack.
+ */
 interface RootComponent {
 
     val rootNavigationStack: Value<ChildStack<*, Child>>
