+++ libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/util/asInstantOrNow.kt
@@ -1,4 +1,4 @@
-package com.github.saikcaskey.pokertracker.data.utils
+package com.github.saikcaskey.pokertracker.domain.util
 
 import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
