+++ libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/util/toUiDateTimeOrNull.kt
@@ -1,4 +1,4 @@
-package com.github.saikcaskey.pokertracker.data.utils
+package com.github.saikcaskey.pokertracker.domain.util
 
 import kotlinx.datetime.Instant
 import kotlinx.datetime.LocalDateTime
